refactor: Age, Fare 패키지 이동
