+++ src/main/java/wooteco/subway/domain/path/Age.java
@@ -1,4 +1,4 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.path;
 
 public class Age {
