+++ src/main/java/wooteco/subway/domain/path/Path.java
@@ -3,8 +3,6 @@ package wooteco.subway.domain.path;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import wooteco.subway.domain.Age;
-import wooteco.subway.domain.Fare;
 import wooteco.subway.domain.Id;
 import wooteco.subway.domain.line.Line;
 import wooteco.subway.domain.line.section.Distance;
